+++ src/main/kotlin/co/makerflow/intellijplugin/state/FlowState.kt
@@ -4,7 +4,6 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.PersistentStateComponent
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
-import com.intellij.openapi.components.service
 import com.intellij.util.xmlb.XmlSerializerUtil
 
 @State(name = "co.makerflow.intellijplugin.state.FlowState", storages = [Storage("MakerflowFlowState.xml")])